Update esm-outline.md
I don't think we should use state as it's a requested changeset. State implies it has knowledge of current state. 

I'm not sure deploy is the right command to add a change request. That would seem to be the final function when the user hits deploy button?

What do you think about adding callbacks from the ECM when you add a call. So the inspector that adds a config change can use callbacks to get notified when the changes are applied?

What about triggering events on the add/remove/commit methods so that we can watch them for changes, say from the deployer bar?